requirements.txt
flask-restplus==0.13.0
Flask-SQLAlchemy==2.1
werkzeug==0.16.0
pytest==5.4.3
requests==2.24.0

venv== API

